fix: validate configs and secrets against the schema before setting
